Publish message after transaction
We were seeing intermittent issues where collection exercises would
fail to be set into `ready for live` state. We observed in the logs that
collection instrument counts from the collection exercise service was 0
after collection instrument had said it have saved a collection
instrument. This was likely because collection exercise processed the
message before the data changes were committed to the database.

This change publishes the message after the data has been saved. The
downside to this is that if a message failed to publish then the data
would not be rolled back. To manage this scenario we log at error level
with relevant information
